clang formatted

+++ search/PathGroup.cc
@@ -1025,8 +1025,8 @@ PathGroups::makeGroupPathEnds(VertexSet &endpoints,
                MakeEndpointPathEnds(visitor, Scene::sceneSet(scenes),
                                     min_max, this));
     for (const auto endpoint : endpoints) {
-      dispatch_queue_->dispatch( [endpoint, &visitors](int i)
-      { visitors[i].visit(endpoint); } );
+      dispatch_queue_->dispatch(
+          [endpoint, &visitors](size_t i) { visitors[i].visit(endpoint); });
     }
     dispatch_queue_->finishTasks();
   }
